refactor: QueryDSL 기반 Repository 구조 변경
