--- order-service/src/main/java/com/palja/order_service/infrastructure/repository/OrderQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.palja.order_service.infrastructure.repository;
-
-public interface OrderQueryDslRepository {
-}
